% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Disease-class}
\alias{Disease-class}
\title{Class "disease"
This class represents the characteristics of the disease}
\description{
Class "disease"
This class represents the characteristics of the disease
}
\section{Slots}{

\describe{
\item{\code{disease}}{disease MeSH id}
}}

