% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease2drug.R
\name{disease2drug,Disease-method}
\alias{disease2drug,Disease-method}
\title{Select the disease related drugs}
\usage{
\S4method{disease2drug}{Disease}(object)
}
\arguments{
\item{diseaseMeshid}{the MeSHid of intersted disease}
}
\value{

}
\description{
Select the disease related drugs
}
