% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease2drug.R
\name{disease2drug}
\alias{disease2drug}
\title{Select the disease related drugs}
\usage{
disease2drug(disease)
}
\arguments{
\item{diseaseMeshid}{the MeSHid of intersted disease}
}
\value{

}
\description{
Select the disease related drugs
}
