% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadDisease2drugDb.R
\name{disease2drugDb}
\alias{disease2drugDb}
\title{This is for load disease2drug.db}
\usage{
disease2drugDb(DEpath)
}
\arguments{
\item{DEpath}{pkgname}
}
\description{
This is for load disease2drug.db
}
