% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease2pathway.R
\name{disease2path}
\alias{disease2path}
\title{Enrichment of pathways associated with drug combinations}
\usage{
disease2path(druglist_meshid)
}
\arguments{
\item{druglist_meshid}{Drug combination of meshid}
}
\value{
data.frame
}
\description{
Enrichment of pathways associated with drug combinations
}
