% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disease2gene.R
\name{drug2gene}
\alias{drug2gene}
\title{Return the disease related genes}
\usage{
drug2gene(drugList)
}
\arguments{
\item{drugList}{List of interest drugs of MeSH id}
}
\description{
Return the disease related genes
}
