% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\title{Enrichment of pathways associated with drug combinations}
\usage{
path(druglist_meshid)
}
\arguments{
\item{druglist_meshid}{Drug combination of meshid}
}
\value{
data.frame
}
\description{
Enrichment of pathways associated with drug combinations
}
